Allow MIDI events mid-frame.
This is mostly done by switching the order of iteration between samples
in the frame and notes in the polyphony, so that we still only update
pitch-bend data and delete released notes once per frame.

The logic for performing the updates loosely resembles midisine.c
(available in this flake as `midisine`) but has the same spirit: we must
have a loose idea of when we can advance to the next event. Rather than
the approach taken in midisine.c, which tries to minimize calls to
jack_midi_event_get(), we lay out a simpler loop which tries to minimize
entering the event-handling logic by keeping it behind a JIT-friendly
conditional.