Rewrite wavetable logic to remove magic numbers.
Many improvements here.

The wavetable is now calibrated so that most notes on my Oxygen 61 are
lerp'd between the upper and lower tables. The conditionals are
presumably still a little expensive and we can do better by clamping the
pitch before passing it to the final lerp. This also gives the lower end
of the wavetable a much richer harmonic count (over 200 at typical
sample rates!) which doesn't appear to be a serious issue now that our
maths are better.

Also, our maths are better! The polynomial has been updated to have much
better coefficients, computed via Remez exchange. The lolremez tool
computed these coefficients; I followed the tutorial at
https://github.com/samhocevar/lolremez/wiki/Tutorial-3-of-5:-changing-variables-for-simpler-polynomials
and renormalized to 1.0. (It turns out that all sines are renormalizable
in this way, so the coefficients are identical to those presented on
that page. More on that in the next commit.)

The table will be recomputed if/when the sample rate changes. This could
be audible but I can't do anything about that. In exchange, we now can
ignore the difference between 44.1 and 48 KHz.