Use an even more accurate fast sine.
We could imagine that our original polynomial is bounded by six
multiplications. Let's view that as six FMAs. A seventh multiplication
used for scaling can be made available here for free. How much better
could we do?

This polynomial is based on the tutorial here:
https://github.com/samhocevar/lolremez/wiki/Tutorial-4-of-5%3A-fixing-lower-order-parameters
It fixes the lowest-order parameter, which is almost 1.0, as exactly
1.0 for faster loading. But it turns out that we instead get an FMA
opportunity because 1.0 premultiplies our input. Our seventh
multiplication and sixth FMA can be arranged as FMA(f(x * x), x, x),
which leaves us with five FMAs for evaluating a polynomial with the
standard unrolling of Horner's rule.

Pass the --debug flag to lolremez along with the given incantation, and
it'll spit out the following maximum relative error: 1.7480848833896522e-11
Doing better with polynomials is probably not possible.

Also, unconditionalize a lerp by clamping the range.